# Documentazione/swagger.yaml
openapi: 3.0.0
info:
  title: WebDeliveryServices API
  description: API RESTful per la gestione di un servizio di food delivery.
  version: 1.0.0
servers:
  - url: /api
    description: Server locale Tomcat
paths:
  /auth/login:
    post:
      summary: Login utente (Op. 1)
      tags:
        - Autenticazione
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              example:
                email: "admin@example.com"
                password: "admin123"
      responses:
        '200':
          description: Login effettuato con successo
          content:
            application/json:
              schema:
                type: object
                example:
                  token: "123e4567-e89b-12d3-a456-426614174000"
                  id_utente: 1
                  nome_completo: "Mario Rossi"
                  ruolo: "admin"
        '400':
          description: Email e password obbligatori
        '401':
          description: Credenziali non valide
        '500':
          description: Errore interno del server
  /auth/logout:
    delete:
      summary: Logout utente (Op. 1)
      tags:
        - Autenticazione
      security:
        - ApiKeyAuth: []
      responses:
        '204':
          description: Logout effettuato con successo (Token rimosso)
        '400':
          description: Token mancante
        '404':
          description: Token non trovato
        '500':
          description: Errore interno del server
  /auth/register:
    post:
      summary: Registrazione Nuovo Cliente (Opzionale)
      tags:
        - Autenticazione
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                nome_completo:
                  type: string
              example:
                email: "cliente@example.com"
                password: "pass"
                nome_completo: "Luigi Verdi"
      responses:
        '201':
          description: Registrazione completata con successo
          content:
            application/json:
              schema:
                type: object
                example:
                  message: "Registrazione completata con successo!"
                  id_utente: 2
        '400':
          description: Parametri obbligatori mancanti
        '409':
          description: Email già in uso
        '500':
          description: Errore interno del server
  /menu:
    get:
      summary: Lettura del menu completo (Op. 2)
      tags:
        - Menu
      responses:
        '200':
          description: Lista dei prodotti con caratteristiche
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
              example:
                - id_prodotto: 1
                  nome: "Pizza Margherita"
                  descrizione: "Pomodoro, mozzarella, basilico"
                  prezzo_base: 6.5
                  caratteristiche:
                    - id_caratteristica: 1
                      nome: "Impasto Integrale"
                      descrizione: "Impasto con farina integrale"
                      differenza_prezzo: 1.5
                      is_default: false
                      id_gruppo: null
        '500':
          description: Errore interno del server

    post:
      summary: Aggiungi prodotto al menu (Extra)
      tags:
        - Menu
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                nome: "Pizza Diavola"
                descrizione: "Pomodoro, mozzarella, salame piccante"
                prezzo_base: 7.5
      responses:
        '201':
          description: Prodotto aggiunto
          content:
            application/json:
              schema:
                type: object
                example:
                  message: "Prodotto aggiunto al menu!"
                  id_prodotto: 5
        '400':
          description: Dati mancanti o invalidi
        '403':
          description: Accesso negato (Solo admin/staff)
        '500':
          description: Errore interno del server
  /menu/search:
    get:
      summary: Ricerca prodotti nel menu (Op. 3)
      tags:
        - Menu
      parameters:
        - in: query
          name: name
          schema:
            type: string
          description: Nome del prodotto
        - in: query
          name: minPrice
          schema:
            type: number
            format: double
          description: Prezzo minimo
        - in: query
          name: maxPrice
          schema:
            type: number
            format: double
          description: Prezzo massimo
      responses:
        '200':
          description: Risultati della ricerca
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
              example:
                - id_prodotto: 1
                  nome: "Pizza Margherita"
                  descrizione: "Pomodoro, mozzarella"
                  prezzo_base: 6.5
        '500':
          description: Errore interno del server
  /menu/{prodId}/ingredients:
    get:
      summary: Lista ingredienti di un prodotto (Op. 10)
      tags:
        - Menu
      parameters:
        - in: path
          name: prodId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ingredienti trovati
          content:
            application/json:
              schema:
                type: object
                example:
                  ingredienti: "Farina, Acqua, Pomodoro, Mozzarella"
        '404':
          description: Prodotto non trovato
        '500':
          description: Errore interno del server
  /menu/{prodId}/features/{featId}:
    delete:
      summary: Eliminazione caratteristica associata a prodotto (Op. 4)
      tags:
        - Menu
      parameters:
        - in: path
          name: prodId
          required: true
          schema:
            type: integer
        - in: path
          name: featId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Caratteristica eliminata con successo
        '404':
          description: Caratteristica non trovata
        '500':
          description: Errore interno del server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization